refactor(roles): extract progression role service
