+++ bot/features/roles/__init__.py
@@ -1,0 +1,5 @@
+"""Progression role feature services."""
+
+from .role_service import ProgressionRoleService
+
+__all__ = ["ProgressionRoleService"]
